requirements.txt
openai
langchain
chromadb
python-dotenv
tiktoken